+++ icekit/page_types/article/models.py
@@ -1,7 +1,8 @@
-from icekit.content_collections.abstract_models import AbstractCollectedContent, \
-    AbstractListingPage, TitleSlugMixin
-from icekit.publishing.models import PublishableFluentContents
 from django.db import models
+from icekit.content_collections.abstract_models import \
+    AbstractCollectedContent, AbstractListingPage, TitleSlugMixin
+from icekit.publishing.models import PublishableFluentContents
+
 
 class ArticleCategoryPage(AbstractListingPage):
     def get_items_to_list(self, request):
@@ -13,12 +14,13 @@ class ArticleCategoryPage(AbstractListingPage):
         return Article.objects.visible().filter(parent_id=unpublished_pk)
 
 
-class Article(PublishableFluentContents, AbstractCollectedContent, TitleSlugMixin):
+class Article(
+        PublishableFluentContents, AbstractCollectedContent, TitleSlugMixin):
     parent = models.ForeignKey(
-        "ArticleCategoryPage",
+        'ArticleCategoryPage',
         limit_choices_to={'publishing_is_draft': True},
         on_delete=models.PROTECT,
     )
 
     class Meta:
-        unique_together = (('parent', 'slug', 'publishing_linked'),)
+        unique_together = (('slug', 'parent', 'publishing_linked'), )