+++ icekit/page_types/article/page_type_plugins.py
@@ -1,6 +1,6 @@
 from fluent_pages.extensions import page_type_pool
-
 from icekit.content_collections.page_type_plugins import ListingPagePlugin
+
 from .models import ArticleCategoryPage