+++ icekit/page_types/layout_page/admin.py
@@ -8,6 +8,6 @@ class UnpublishableLayoutPageAdmin(FluentLayoutsMixin, FluentContentsPageAdmin):
     raw_id_fields = ('parent',)
 
 
-class LayoutPageAdmin(FluentLayoutsMixin, FluentContentsPageAdmin,
-                      PublishingAdmin):
+class LayoutPageAdmin(
+        FluentLayoutsMixin, FluentContentsPageAdmin, PublishingAdmin):
     raw_id_fields = ('parent',)