+++ icekit/project/settings/glamkit.py
@@ -1,0 +1,11 @@
+from .icekit import *
+
+# DJANGO ######################################################################
+
+INSTALLED_APPS += (
+    'sponsors',
+
+    'icekit_events',
+    'icekit_events.event_types.simple',
+    'icekit_events.page_types.eventlistingfordate',
+)